+++ R/playwidget.R
@@ -3,6 +3,7 @@
 #'
 #' @export
 playwidgetOutput <- function(outputId, width = '0px', height = '0px') {
+	registerShinyHandlers()
   shinyWidgetOutput(outputId, 'rglPlayer', width, height, package = 'rgl')
 }
 
@@ -10,8 +11,9 @@ playwidgetOutput <- function(outputId, width = '0px', height = '0px') {
 #'
 #' @export
 renderPlaywidget <- function(expr, env = parent.frame(), quoted = FALSE, outputArgs = list()) {
+	registerShinyHandlers()
   if (!quoted) expr <- substitute(expr)  # force quoted
-  markRenderFunction(playwidgetOutput,
+  shiny::markRenderFunction(playwidgetOutput,
                      shinyRenderWidget(expr, playwidgetOutput, env, quoted = TRUE),
   		     outputArgs = outputArgs)
 }