+++ R/rglwidget.R
@@ -1,6 +1,3 @@
-# Shiny objects if the widget sets elementId, so we
-# need to detect it.  Thanks to Joe Cheng for suggesting this code.
-inShiny <- function() !is.null(getDefaultReactiveDomain())
 
 rmarkdownOutput <- function() {
   if (requireNamespace("rmarkdown", quietly = TRUE)) {
@@ -364,95 +361,6 @@ rglwidget <- local({
   } else
     result
 }})
-
-#' Widget output function for use in Shiny
-#'
-#' @export
-rglwidgetOutput <- function(outputId, width = '512px', height = '512px') {
-  shinyWidgetOutput(outputId, 'rglWebGL', width, height, package = 'rgl')
-}
-
-#' Widget render function for use in Shiny
-#'
-#' @export
-renderRglwidget <- function(expr, env = parent.frame(), quoted = FALSE, outputArgs = list()) {
-  if (!quoted) expr <- substitute(expr)  # force quoted
-  markRenderFunction(rglwidgetOutput,
-                     shinyRenderWidget(expr, rglwidgetOutput, env, quoted = TRUE),
-  		     outputArgs = outputArgs)
-}
-
-shinySetPar3d <- function(..., session,
-                          subscene = currentSubscene3d(cur3d())) {
-  if (!requireNamespace("shiny", quietly = TRUE))
-    stop("function requires shiny")
-  args <- list(...)
-  argnames <- names(args)
-  if (length(args) == 1 && is.null(argnames)) {
-    args <- args[[1]]
-  }
-  # We might have been passed modified shinyGetPar3d output;
-  # clean it up.
-  args[["subscene"]] <- NULL
-  args[["tag"]] <- NULL
-  argnames <- names(args)
-  
-  if (is.null(argnames) || any(argnames == ""))
-    stop("Parameters must all be named")
-  
-  badargs <- argnames[!(argnames %in% .Par3d) | argnames %in% .Par3d.readonly]
-  if (length(badargs))
-    stop("Invalid parameter(s): ", badargs)
-  
-  for (arg in argnames) {
-    session$sendCustomMessage("shinySetPar3d", 
-                              list(subscene = subscene,
-                                   parameter = arg,
-                                   value = args[[arg]]))
-  }
-}
-
-shinyGetPar3d <- function(parameters, session,
-                          subscene = currentSubscene3d(cur3d()),
-                          tag = "") {
-  badargs <- parameters[!(parameters %in% .Par3d)]
-  if (length(badargs))
-    stop("invalid parameter(s): ", badargs)
-  session$sendCustomMessage("shinyGetPar3d",
-                              list(tag = tag, subscene = subscene, 
-                                   parameters = parameters))
-}
-
-convertShinyPar3d <- function(par3d, ...) {
-  for (parm in c("modelMatrix", "projMatrix", "userMatrix", "userProjection"))
-    if (!is.null(par3d[[parm]]))
-      par3d[[parm]] <- matrix(unlist(par3d[[parm]]), 4, 4)
-  for (parm in c("mouseMode", "observer", "scale", "viewport", "bbox", "windowRect"))
-    if (!is.null(par3d[[parm]]))
-      par3d[[parm]] <- unlist(par3d[[parm]])
-  par3d
-}
-
-shinyResetBrush <- function(session, brush) {
-  session$sendCustomMessage("resetBrush", brush)
-}
-
-convertShinyMouse3d <- function(mouse3d, ...) {
-  for (parm in c("model", "proj")) 
-    if (!is.null(mouse3d[[parm]]))
-      mouse3d[[parm]] <- matrix(unlist(mouse3d[[parm]]), 4, 4)
-    
-  if (length(unlist(mouse3d$view)) == 4) 
-    mouse3d$view <- structure(unlist(mouse3d$view),
-                              names = c("x", "y", "width", "height"))
-  if (all(c("p1", "p2") %in% names(mouse3d$region)))
-    mouse3d$region <- with(mouse3d$region,
-                           c(x1 = (p1$x + 1)/2,
-                             y1 = (p1$y + 1)/2,
-                             x2 = (p2$x + 1)/2,
-                             y2 = (p2$y + 1)/2))
-  structure(mouse3d, class = "rglMouseSelection")
-}
 
 print.rglMouseSelection <- function(x, verbose = FALSE, ...) {
   if (!is.null(x$region)) {